// src/app/main-home/providers/models/service.interface.ts
export interface ServiceI {
    idServicios: number;
    nombreServicio: string;
    descripcionServicio: string;
    precioEstimado: number;
    imagen: string;
    historialEmpresas: string;
    numeroContacto: string;
    correoContacto: string;
    estadoServicio: number;
    Proveedor_idProveedor: number;
    TipoServicio_idtipoServicio: number;
}